requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
supabase>=2.3.0
pydantic>=2.6.0
pandas>=2.2.0
python-dateutil>=2.8.2